Add typescript, adjust types, and also adjust README.md
