+++ examples/react/postprocessing/src/App.jsx
@@ -1,7 +1,10 @@
 /* eslint-disable react/no-unknown-property */
-import { Canvas } from '@react-three/fiber'
+import { Canvas, extend } from '@react-three/fiber'
 import Effect from "./component/postprocessing"
 import './App.css'
+import { OrbitControls } from 'three-stdlib'
+
+extend({ OrbitControls });
 
 function App() {
