--- package/src/react-modules/helper/index.ts
@@ -1,2 +1,0 @@
-export * from "./ts-utils"
-export * from "./types"
