+++ package/src/react-modules/index.tsx
@@ -1,8 +1,5 @@
 // Component
 export * from './component'
 
-// Helper
-export * from "./helper"
-
 // Hooks
 export * from "./hooks"
