+++ package/src/types-helper/index.ts
@@ -1,0 +1,1 @@
+export * from "./ts-utils"
