+++ package/src/vanilla-modules/index.ts
@@ -1,5 +1,2 @@
-// Vanilla Types
-export * from "./types"
-
 // Main Class
 export * from "./vanilla"
