+++ rollup.config.js
@@ -50,6 +50,19 @@ export default [
         plugins: [babel(getBabelOptions({ useESModules: false })), resolve({ extensions })]
     },
     {
+        input: `./src/types.ts`,
+        output: { file: `dist/types.js`, format: 'esm', exports: 'auto' },
+        external,
+        plugins: [babel(getBabelOptions({ useESModules: true }, '>1%, not dead, not ie 11, not op_mini all')), resolve({ extensions })],
+        preserveModules: true
+    },
+    {
+        input: `./src/types.ts`,
+        output: { file: `dist/types.cjs.js`, format: 'cjs', exports: 'auto' },
+        external,
+        plugins: [babel(getBabelOptions({ useESModules: false })), resolve({ extensions })]
+    },
+    {
         input: `./src/r3f.ts`,
         output: { file: `dist/r3f.js`, format: 'esm', exports: 'auto'},
         external,
